project1/requirements.txt
numpy
pandas
seaborn
matplotlib
scikit-learn
xgboost
catboost
dill
Flask
psycopg2
mysql-connector-python

# -e .
